requirements.txt
altair==5.5.0
attrs==25.3.0
bcrypt==4.1.2
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.0.2
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysql-connector-python==8.3.0
narwhals==1.48.0
numpy==1.26.4
packaging==24.2
pandas==2.2.2
pillow==10.4.0
plotly==5.21.0
protobuf==4.25.8
pyarrow==21.0.0
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rich==13.9.4
rpds-py==0.26.0
six==1.17.0
smmap==5.0.2
streamlit==1.48.1
streamlit-javascript==0.1.5
tenacity==8.5.0
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.3
